Refactor statistical analysis table
